Refactor IPay88 to follow correct integration using Notification
